requirements.txt
et-xmlfile==1.1.0
lxml==4.9.2
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
six==1.16.0
